requirements.txt
PyYAML >=6.0, ==6.*
discord.py >=2.3.1, ==2.3.*
typer >=0.9.0, ==0.9.*
